Add autocompletion for usernames (#15)
* Add autocompletion for usernames

This progressively-enhances the kudos entry page for clients with JavaScript
enabled. It replaces the <select> of usernames with a text field that presents
possible completion matches when they start typing. The user can click to select
a username and this updates the <select> behind the scenes so all the form
submission machinery still just works.

Closes #11.

* Clear the underlying <select> if clear selection

* couple tweaks

* mark this spec pending